+++ R/cnd.R
@@ -279,63 +279,132 @@ muffle <- function(...) NULL
 
 #' Signal an error, warning, or message
 #'
+#' @description
+#'
 #' These functions are equivalent to base functions [base::stop()],
-#' [base::warning()] and [base::message()], but the `type` argument
-#' makes it easy to create subclassed conditions. They also don't
-#' include call information by default. This saves you from typing
-#' `call. = FALSE` to make error messages cleaner within package
-#' functions.
+#' [base::warning()] and [base::message()], but make it easy to supply
+#' condition metadata:
 #'
-#' Like `stop()` and [cnd_abort()], `abort()` signals a critical
-#' condition and interrupts execution by jumping to top level (see
-#' [rst_abort()]). Only a handler of the relevant type can prevent
-#' this jump by making another jump to a different target on the stack
-#' (see [with_handlers()]).
+#' * Supply `type` to create a classed condition. Typed conditions can
+#'   be captured or handled selectively, allowing for finer-grained
+#'   error handling.
 #'
-#' `warn()` and `inform()` both have the side effect of displaying a
-#' message. These messages will not be displayed if a handler
-#' transfers control. Transfer can be achieved by establishing an
-#' exiting handler that transfers control to [with_handlers()]). In
-#' this case, the current function stops and execution resumes at the
-#' point where handlers were established.
+#' * Supply metadata with named `...` arguments. This data will be
+#'   stored in the condition object and can be examined by handlers.
 #'
-#' Since it is often desirable to continue normally after a message or
-#' warning, both `warn()` and `inform()` (and their base R equivalent)
-#' establish a muffle restart where handlers can jump to prevent the
-#' message from being displayed. Execution resumes normally after
-#' that. See [rst_muffle()] to jump to a muffling restart, and the
-#' `muffle` argument of [inplace()] for creating a muffling handler.
 #'
-#' @param msg A message to display.
-#' @param type Subclass of the condition to signal.
-#' @param call Whether to display the call. If a number `n`, the call
+#' @section Call trace:
+#'
+#' Unlike `stop()` and `warning()`, these functions don't include call
+#' information by default. This saves you from typing `call. = FALSE`
+#' and produces cleaner error messages.
+#'
+#'
+#' @section Mufflable conditions:
+#'
+#' Signalling a condition with `inform()` or `warn()` causes a message
+#' to be displayed in the console. These messages can be muffled with
+#' [base::suppressMessages()] or [base::suppressWarnings()].
+#'
+#'
+#' @section Lifecycle:
+#'
+#' These functions were changed in rlang 0.3.0 to take condition
+#' metadata with `...`. Consequently:
+#'
+#' * All arguments were renamed to be prefixed with a dot.
+#' * `.call` (previously `call`) can no longer be passed positionally.
+#'
+#' @param .msg The message to display.
+#' @param .type Subclass of the condition.
+#' @param ... Additional data to be stored in the condition object.
+#' @param .call Whether to display the call. If a number `n`, the call
 #'   is taken from the nth frame on the [call stack][call_stack].
+#' @param msg,type,call These arguments were renamed to be prefixed
+#'   with a dot and are deprecated as of rlang 0.3.0.
+#'
 #' @export
-abort <- function(msg, type = NULL, call = FALSE) {
-  cnd <- error_cnd(type, .msg = msg, .call = cnd_call(call))
-  if (!is_false(call)) {
+#' @examples
+#' # These examples are guarded to avoid throwing errors
+#' if (FALSE) {
+#'
+#' # Signal an error with a message just like stop():
+#' abort("Something bad happened")
+#'
+#' # Give a class to the error:
+#' abort("Something bad happened", "somepkg_bad_error")
+#'
+#' # This will allow your users to handle the error selectively
+#' tryCatch(
+#'   somepkg_function(),
+#'   somepkg_bad_error = function(err) {
+#'     cnd_warn(err) # Demote the error to a warning
+#'     NA            # Return an alternative value
+#'   }
+#' )
+#'
+#' # You can also specify metadata that will be stored in the condition:
+#' abort("Something bad happened", "somepkg_bad_error", data = 1:10)
+#'
+#' # This data can then be consulted by user handlers:
+#' tryCatch(
+#'   somepkg_function(),
+#'   somepkg_bad_error = function(err) {
+#'     # Compute an alternative return value with the data:
+#'     recover_error(err$data)
+#'   }
+#' )
+#'
+#' }
+abort <- function(.msg, .type = NULL, ..., .call = FALSE, msg, type, call) {
+  validate_signal_args(msg, type, call)
+
+  cnd <- error_cnd(.type, ..., .msg = .msg, .call = cnd_call(.call))
+  if (!is_false(.call)) {
     cnd$call <- cnd$.call
   }
+
   stop(cnd)
 }
 #' @rdname abort
 #' @export
-warn <- function(msg, type = NULL, call = FALSE) {
-  cnd <- warning_cnd(type, .msg = msg, .call = cnd_call(call))
-  if (!is_false(call)) {
+warn <- function(.msg, .type = NULL, ..., .call = FALSE, msg, type, call) {
+  validate_signal_args(msg, type, call)
+
+  cnd <- warning_cnd(.type, ..., .msg = .msg, .call = cnd_call(.call))
+  if (!is_false(.call)) {
     cnd$call <- cnd$.call
   }
+
   warning(cnd)
 }
 #' @rdname abort
 #' @export
-inform <- function(msg, type = NULL, call = FALSE) {
-  msg <- paste0(msg, "\n")
-  cnd <- message_cnd(type, .msg = msg, .call = cnd_call(call))
-  if (!is_false(call)) {
+inform <- function(.msg, .type = NULL, ..., .call = FALSE, msg, type, call) {
+  validate_signal_args(msg, type, call)
+
+  .msg <- paste0(.msg, "\n")
+  cnd <- message_cnd(.type, ..., .msg = .msg, .call = cnd_call(.call))
+  if (!is_false(.call)) {
     cnd$call <- cnd$.call
   }
+
   message(cnd)
+}
+
+validate_signal_args <- function(msg, type, call, env = parent.frame()) {
+  if (!missing(msg)) {
+    warn("`msg` has been renamed to `.msg` and is deprecated as of rlang 0.3.0")
+    env$.msg <- msg
+  }
+  if (!missing(type)) {
+    warn("`type` has been renamed to `.type` and is deprecated as of rlang 0.3.0")
+    env$.type <- type
+  }
+  if (!missing(call)) {
+    warn("`call` has been renamed to `.call` and is deprecated as of rlang 0.3.0")
+    env$.call <- call
+  }
 }
 
 #' Catch a condition